Reorganise backend
