+++ Server/Repositories/ITrelloKillerRepository.cs
@@ -1,0 +1,7 @@
+﻿namespace aspnetreact.Server.Repositories
+{
+    public interface ITrelloKillerRepository
+    {
+        void AddList(TrelloKillerList trelloKillerList);
+    }
+}
