+++ Startup.cs
@@ -1,4 +1,5 @@
-using aspnetreact.Server;
+using aspnetreact.Server.DomainServices;
+using aspnetreact.Server.Repositories;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.SpaServices.Webpack;
@@ -21,6 +22,7 @@ namespace aspnetreact
         {
             services.AddMvc();
             services.AddSingleton<IDatabaseConnection, DatabaseConnection>();
+            services.AddSingleton<ITrelloKillerRepository, TrelloKillerRepository>();
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
